Modelling growth and its uncertainty in FLa4a
This was mainly cleaning up the text. There was one place (associated
with the copula example) that said “Needs more explanation”, but I have
removed it as I think running the plots are self-explanatory, and this
example already does what preceding examples do (but just implements a
different multivariate statistical distribution for the var-cov matrix).
Copulas are quite complicated and I don’t think here is the place to
explain them.