Natural_mortality_modelling_in_FLa4a
Cleaning up of the text and minor edits. In one example, I made the use
of a CV and correlation in setting up the variance-covariance matrix
clearer, with the actual correlation correctly specified (it was
approximated before).

Same issues as growth modelling in Fla4a:
• Internet access required for NAO example, so I have added a warning
sentence.
• Currently not sure how to include the automated Figure numbers in the
Figure Captions to make referencing clearer.

Reference to Kenshington 2013 missing.

TO DO:
• There is a note asking how ageing error can be included (at the end of
the first “Adding uncertainty…” section).
• The following appears when discussing the m() method for the first
time:
The method uses the `range` slot to work out the dimensions of the
`FLQuant` object.
% Future developments will also allow for easy insertion into FLStockLen
objects.
I have left these in for now.