Forecasting_on_the_Medium_Term_for_advice_using_FLasher
This is by far the longest tutorial, and the most problematic!

This tutorial will not be functional yet for general users, for the
following reasons:
• The FLasher package is not yet available to download via
“install.packages(c("FLasher"), repos="http://flr-project.org/R")”
The message one gets when attempting this is
package ‘FLasher’ is not available (for R version 3.3.x)
In my case, x=2, but I think this applies to later versions too. In
order to get the tutorial to work, I used a local copy of the FLasher
package (FLasher_0.0.2.9006.zip)
• The tutorial contains setMethod code for stf, which is hidden from the
user who follows, e.g. the html version of the tutorial, so the tutorial
may not work for them.

There is a serious problem with Example 4. The control object:
ctrl_rel_catch <- fwdControl(
data.frame(year = 2009:2018,
quant = "catch",
value = 0.9,
relYear = 2008:2017))
is meant to set catches in 2009-2018 at 90% of the catches in the
immediately preceding year, but it seems the value of 0.9 does not do
the job and needs to be multiplied by 10 (then it works as expected). So
to get a constant catch, you need to set value=10, and to get 90% as
this example intended, you need to set value=9, which does not make
intuitive sense (it is neither 0.9 nor 90%). I have changed it, for now,
to 9 to do as intended, but this needs to be fixed.

There is also a serious problem with Example 5. In the control object:
df <- data.frame(
year = rep(2009:2018, each=2),
quant = c("f","catch"),
value = c(f01, NA),
min = c(NA, min_catch))
Both f01 and min_catch need to be multiplied by 20 to get the desired
result. Strange! I have included this multiplication in the control
object, but have warned of the bug.

Example 6 is also not working. At this point, I abandoned the tutorial,
as it is not possible to check the examples. In fact, looking closely at
Examples 1-3, they are not working either. For example 1, f_status_quo
needs to be multiplied by 10 to get the desired result. For example 2,
future_catch needs to be multiplied by 10 for the desired result.
Finally, for the final section of example 3 when a longer projection
period is tried, future_ssb needs to be multiplied by 10 for the desired
result.

As with other tutorials
• Need to improve cross-referencing to other tutorials (e.g. through
“live” links). Here it is to STF and Stock-recruit fitting.
• Need to improve auto-numbering of Figures to the Figure captions so
that it is clear which Figure number (referenced in text) belongs to
which Figure.