Loading_your_data_into_FLR
1. There was a path issue when reading data, e.g.:
catch.n <- read.csv("src/catch_numbers.csv",row=1)
This does not work for somebody completing the tutorial but without the
data pre-loaded. A solution could have been:
catch.n <-
read.csv("http://flr-project.org/doc/src/catch_numbers.csv",row=1)
Although this works for read.csv and readFLIndices, it does not work
with readFLStock or readVPAFile, because these don’t function with URLs.
The solution (provided by Iago) was to download the zip file, either to
a temporary folder or to a folder selected by the user, and then to
unzip it and work from there.
2. There is no help when typing ?readVPAFile, ?readFLStock, or
?readFLIndices.
3. I have removed all @ usages, and instead used methods [e.g.
catch(xxx) instead of xxx@catch].
4. The final example named “GeneratingFLFleets” is not functional.