Plotting_FLR_objects_using_lattice
1. plot(nsher) does not seem to work properly
2. Included Sarkar 2008 reference
3. Minor edits throughout.
4. Question about whether this statement is relevant “To plot FLR
objects with lattice a conversion to data.frame is required to select
columns names for dims and FLR objects (‘data’) to plot. It is usefull
to check how the objects look like before plotting it”. This is not
obvious in the examples given. Maybe the method converts FLQuants into
data.frame.