Reference_points_for_fisheries_management_with_FLBRP
1. This tutorial is like a mini-thesis to start off with, providing a
fairly extensive introduction to reference points – this is quite
different to the other tutorials
2. Referencing needs sorting out within this tutorial and between
tutorials. For example, this tutorial makes some use of
“bibliography.bib”, but there are references that are not included in
this file (e.g. Doulman 1995). Making use of this file with e.g.
@fao1996 means that references go right at the end of the tutorial,
which seems a strange place to put them – not sure how to arrange it so
that they go early, e.g. before “more information”
3. NO HELP PAGES FOR FLBRP! The link http://www.flr-project.org/FLBRP/
does not work. This is a serious shortcoming.
4. The link to other tutorials is not working: extract from this
tutorial: “(see the tutorial on [modelling stock recruitment with
FLSR](http://doc.flr-project.org/Modelling_stock_recruitment_with_FLSR.html)”
– this does not work.
5. Changed reference to “crash” to “Fcrash” because it is the f that
leads to a crash, although refpts() still show it is “crash”.
6. “Yield” was not clearly defined, and it was confusing seeing “yield”
observations in plot(brp4Ri,obs=T), so the following text was added as a
preamble to the plot:
Corresponding assessment estimates of SSB, recruitment and F, as well as
observed landings representing "yield" can easily be added to these
plots. It should be noted that the curves and associated reference
points in these plots represent equilibrium states, whereas the
assessment estimates and observations are non-equilibrium. Furthermore,
F will always represent total catch (landings and discards), whereas the
"yield" is only that portion of the catch that one would desire to
optimise (i.e. landings) - so effectively one is looking for the F
(total catch) that will optimise the yield (landings only) when
estimating MSY reference points.
7. The price and costs of the economic reference point example had to be
adjusted so that the FMEY was lower than the FMSY, to fit the story in
the example!
8. Otherwise minor edits.